fix: keep world transforms in double precision for CPU geometry
The graphics world transform was stored as Matrix4f, so CPU-side geometry (bounding boxes, picking) was computed from single-precision values. A rotation round-tripped through float carries an ~1e-7 non-orthogonality residual (R·R^T - I), which the double-epsilon-scale checks in `BoundingBox::Rotate` rejected, throwing during otherwise valid bounding-box / picking computations. Hold the transform in double as the canonical value and narrow to float only at the GPU boundary.

Transform precision:
- `IEntityGraphics::world_transform_` and `SetWorldTransform` are now `Matrix4d`; child propagation in `EntityGraphics` composes in double
- Add `GetWorldTransformD()` returning the double canonical value; `GetWorldTransform()` keeps returning `Matrix4f` for GPU upload by narrowing the double value (via a named local to avoid a dangling reference to Eigen's lazy cast expression)
- `EntityRenderer::WalkAssembly` sets the accumulated transform in double; picking/selection call sites (`PointGraphics`, `CopiousDataGraphics`, `TrimmedSurfaceGraphics`) and the bounding-box paths now read the double value directly instead of casting float back to double

Float-aware rotation tolerance:
- Add `kFloatGeometryTolerance` (1e-5) for geometry that passed through single precision, plus `IsRotation` (orthogonal + det +1) and `NearestRotation` (Gram-Schmidt re-orthonormalization, right-handed) in numerics/tolerance
- `BoundingBox::Rotate` validates with `IsRotation` (float scale, rejecting only genuine scale/shear/NaN inputs) and normalizes the matrix via `NearestRotation` before applying, preventing residual accumulation
- Guard `GetWorldBoundingBox` in the picking path: a degenerate rotation component returns nullopt instead of throwing across the const pick path, falling back to depth estimation

Add tests for `IsRotation` / `NearestRotation` and `BoundingBox::Rotate` under near-rotation (float-residual) inputs.

+++ src/graphics/core/i_entity_graphics.cpp
@@ -80,7 +80,7 @@ void IEntityGraphics::SetGlobalParam(
     global_param_ = copy;
 }
 
-void IEntityGraphics::SetWorldTransform(const igesio::Matrix4f& matrix) {
+void IEntityGraphics::SetWorldTransform(const igesio::Matrix4d& matrix) {
     world_transform_ = matrix;
 }
